#381: move test module to resources
